Move DetailsView and HomeView to src/screen
