+++ nonCisMaleInCs/src/screens/HomeView.js
@@ -1,0 +1,28 @@
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * @format
+ * @flow strict-local
+ */
+
+import React from 'react';
+import {StyleSheet, View, Text} from 'react-native';
+
+function HomeView() {
+  return (
+    <View style={styles.homeView}>
+      <Text>Home View!</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  homeView: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
+export default HomeView;
